requirements.txt

fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
httpx==0.26.0
python-dotenv==1.0.1
openai==1.12.0

# Optional but recommended for production
gunicorn==21.2.0
python-multipart==0.0.9